Remove Category model, replacing with parent/child relationships on the Forum class.
TODO: Top-level forums should be categories, but for now we have an explicit is_category boolean - this is because I'm not sure how to get the 'parent' form to contain a "Null" option.  Once we can get a "Null" option in the parents list, we can replace the is_category boolean code with "parent is None" checks.