Apply sorting on subsequent pages
Modified render_pagination macro by adding two optional arguemts sort_by and asc.
If sorted is set as a string, sort_by and order_by (asc if asc set to True,
desc otherwise) are injected to the page links.
This helps to keep the sorting of the members in memberlist if the users wants to see
e.g. the second page of users with most comments