Improve speed for bulk inserts
It has a little flaw though.
The last_post_id is also the first_post_id. This is because otherwise
it would be around 10x slower (but still way faster than before).
It doesn't impact the other stuff at all and it is just
being used for testing purposes anyway.