Redirect user to first unread post of a topic
If there are new posts to a topic the users usually wants to continue
reading at the first post (s)he has not read.
Therefore the users should be redirected to the first unread post when
clicking the topic in the forum view. If there is no unread post, the
user will be send to the topic itself, i.e. the first post