[BUGFIX] Make `flaskbb install` work with MySQL (#268)
* [BUGFIX] Make `flaskbb install` work with MySQL

The install script offers the choice to drop an existing database first.
It then proceeds with the `upgrade` procedure. This doesn't work
(at least) for MySQL. Processing stops with an "Database not found"
error. So let's create the database first if it's missing and then
continue with upgrading.
With this change installation works for me (MariaDB 10.1).