Take territory code into account during language selection
This fixes (#299) a bug that made translations with territory code
'unsavable'. The reason for this is, that the Locale.language attribute
returns the language without the territory code. Translations like
'de_AT' or 'pt_BR' where unusable because of that -- 'de' or 'pt' would
work because no territory code (= AT, BR, etc) is used.